pyproject.toml
[project]
name = "asocks-api"
version = "0.1.0"
description = "Python client for interacting with the asocks.com API"
authors = [
    {name = "Platon U.",email = "revlze@yahoo.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [{ include = "asocks_api" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.2.1"

